+++ web-app/public/locales/en/common.json
@@ -2,7 +2,7 @@
   "add": "Add",
   "delete": "Delete",
   "error": "Error",
-  "close": "Close",
+  "back": "Back",
   "reset": "Reset",
   "invoice": {
     "supplier": "Supplier",
@@ -48,14 +48,18 @@
   "download": "Download",
   "loadMore": "Load more",
   "invoiceDocumentation": "Invoice documentation",
-  "auth": {
+  "unauthorized": {
     "title": "Unauthorized access",
     "description": "You have no permission to view this page",
     "upvsLogout": "Logout from slovensko.sk"
   },
-  "notfound": {
-    "title": "Not found",
-    "description": "Page not found"
+  "notFound": {
+    "title": "Page not found",
+    "description": [
+      "If you typed the web address, check it is correct.",
+      "If you pasted the web address, check you copied the entire address.",
+      "If the web address is correct or you selected a link or button, write an email to"
+    ]
   },
   "loading": {
     "title": "Processing request"
