+++ web-app/public/locales/sk/common.json
@@ -2,7 +2,7 @@
   "add": "Pridať",
   "delete": "Odstrániť",
   "error": "Chyba",
-  "close": "Zavrieť",
+  "back": "Naspäť",
   "reset": "Vynulovať",
   "invoice": {
     "supplier": "Dodávateľ",
@@ -48,14 +48,18 @@
   "download": "Stiahnuť",
   "loadMore": "Načítaj viac",
   "invoiceDocumentation": "Dokumentácia faktúr",
-  "auth": {
+  "unauthorized": {
     "title": "Nepovolený prístup",
     "description": "Nemáte povolenie na prezeranie tejto stránky",
     "upvsLogout": "Odhlásiť zo slovensko.sk"
   },
-  "notfound": {
-    "title": "Nenájdené",
-    "description": "Stránka nebola nájdená"
+  "notFound": {
+    "title": "Stránka nebola nájdená",
+    "description": [
+      "Ak ste napísali webovú adresu, skontrolujte, či je správna.",
+      "Ak ste skopírovali webovú adresu, skontrolujte, či ste skopírovali celú adresu.",
+      "Ak je webová adresa správna alebo ste sa tu dostali cez odkaz alebo tlačidlo, napíšte email na adresu"
+    ]
   },
   "loading": {
     "title": "Spracúvam žiadosť"
