+++ web-app/src/InvoiceList/index.js
@@ -1,7 +1,7 @@
 import {useCallback} from 'react'
 import {useDispatch, useSelector} from 'react-redux'
 import {Link, useLocation} from 'react-router-dom'
-import {Button, Card, Table} from 'react-bootstrap'
+import {Button, Table} from 'react-bootstrap'
 import {useTranslation} from 'react-i18next'
 import classnames from 'classnames'
 import {get} from 'lodash'
@@ -22,45 +22,43 @@ export default ({getInvoicesAction, path, title}) => {
   )
 
   return (
-    <Card className="m-1">
-      <Card.Header className="bg-primary text-white text-center" as="h3">{title}</Card.Header>
-      <Card.Body>
-        <Filters getInvoices={getInvoices} />
-        {invoiceIds && <>
-          <Table striped hover responsive size="sm">
-            <thead>
-              <tr>
-                <th>#</th>
-                <th>{t('invoice.supplierIco')}</th>
-                <th>{t('invoice.customerIco')}</th>
-                <th>{t('invoice.issueDate')}</th>
-                <th>{t('invoice.amount')}</th>
-                <th>{t('invoice.format')}</th>
-                <th />
+    <>
+      <h1 className="govuk-heading-l">{title}</h1>
+      <Filters getInvoices={getInvoices} />
+      {invoiceIds && <>
+        <Table striped hover responsive size="sm">
+          <thead>
+            <tr>
+              <th>#</th>
+              <th>{t('invoice.supplierIco')}</th>
+              <th>{t('invoice.customerIco')}</th>
+              <th>{t('invoice.issueDate')}</th>
+              <th>{t('invoice.amount')}</th>
+              <th>{t('invoice.format')}</th>
+              <th />
+            </tr>
+          </thead>
+          <tbody>
+            {invoiceIds.map((invoiceId, i) => (
+              <tr key={i} className={classnames({'text-secondary': invoices[invoiceId].test})}>
+                <td>{i + 1}</td>
+                <td>{invoices[invoiceId].supplierIco}</td>
+                <td>{invoices[invoiceId].customerIco}</td>
+                <td>{invoices[invoiceId].issueDate}</td>
+                <td>{invoices[invoiceId].amount}</td>
+                <td>{invoices[invoiceId].format}</td>
+                <td>
+                  <Link to={`/invoices/${invoiceId}`}>detail</Link>
+                </td>
               </tr>
-            </thead>
-            <tbody>
-              {invoiceIds.map((invoiceId, i) => (
-                <tr key={i} className={classnames({'text-secondary': invoices[invoiceId].test})}>
-                  <td>{i + 1}</td>
-                  <td>{invoices[invoiceId].supplierIco}</td>
-                  <td>{invoices[invoiceId].customerIco}</td>
-                  <td>{invoices[invoiceId].issueDate}</td>
-                  <td>{invoices[invoiceId].amount}</td>
-                  <td>{invoices[invoiceId].format}</td>
-                  <td>
-                    <Link to={`/invoices/${invoiceId}`}>detail</Link>
-                  </td>
-                </tr>
-              ))}
-            </tbody>
-          </Table>
-          {nextId &&
-            <Button variant="outline-primary" block onClick={() => getInvoices(search, nextId)}>
-              {t('loadMore')}
-            </Button>}
-        </>}
-      </Card.Body>
-    </Card>
+            ))}
+          </tbody>
+        </Table>
+        {nextId &&
+          <Button variant="outline-primary" block onClick={() => getInvoices(search, nextId)}>
+            {t('loadMore')}
+          </Button>}
+      </>}
+    </>
   )
 }
