+++ web-app/src/helpers/Unauthorized.js
@@ -3,10 +3,15 @@ import {useTranslation} from 'react-i18next'
 export default () => {
   const {t} = useTranslation('common')
   return (
-    <div>
-      <h1>401</h1>
-      <h2>{t('auth.title')}</h2>
-      <div>{t('auth.description')}</div>
-    </div>
+    <main className="govuk-main-wrapper govuk-main-wrapper--l" id="main-content" role="main">
+      <div className="govuk-grid-row">
+        <div className="govuk-grid-column-two-thirds">
+          <h1 className="govuk-heading-l">{t('unauthorized.title')}</h1>
+          <p className="govuk-body">
+            {t('unauthorized.description')}
+          </p>
+        </div>
+      </div>
+    </main>
   )
 }
